+++ src/application/modules/subscription.module.ts
@@ -4,9 +4,9 @@ import { PrismaService } from 'src/infrastructure/prisma/prisma.service';
 import { WeatherModule } from 'src/application/modules/weather.module';
 import { EmailModule } from './email.module';
 import { TokenModule } from './token.module';
-import { SubscriptionController } from 'src/presentation/subscription.controller';
+import { SubscriptionController } from '../../presentation/subscription/subscription.controller';
 import { SubscriptionRepository } from 'src/infrastructure/subscription/subscription.repository';
-import { SubscriptionRepositoryToken } from 'src/core/subscription/subscription-repoository.interface';
+import { SubscriptionRepositoryToken } from 'src/application/subscription/interfaces/subscription-repoository.interface';
 
 @Module({
   imports: [EmailModule, TokenModule, WeatherModule],
