+++ src/common/pipes/city-validation.pipe.ts
@@ -1,10 +1,10 @@
 import { PipeTransform, Injectable, Inject } from '@nestjs/common';
-import { CreateSubscriptionDto } from '../../application/subscription/dto/create-subscription.dto';
+import { CreateSubscriptionDto } from '../../presentation/subscription/dto/create-subscription.dto';
 import {
   CityValidatable,
   WeatherClientToken,
 } from '../../core/weather/weather.interface';
-import { GetWeatherDto } from 'src/application/weather/dto/get-weather.dto';
+import { GetWeatherDto } from 'src/presentation/weather/dto/get-weather.dto';
 
 @Injectable()
 export class CityValidationPipe implements PipeTransform {
