+++ src/presentation/subscription/subscription.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Post, Body, Get, Param, UsePipes } from '@nestjs/common';
-import { SubscriptionService } from '../application/subscription/use-cases/subscription.service';
-import { CreateSubscriptionDto } from '../application/subscription/dto/create-subscription.dto';
+import { SubscriptionService } from 'src/application/subscription/use-cases/subscription.service';
+import { CreateSubscriptionDto } from './dto/create-subscription.dto';
 import { CityValidationPipe } from 'src/common/pipes/city-validation.pipe';
 
 @Controller()
