+++ src/presentation/weather/weather.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Query, UsePipes } from '@nestjs/common';
-import { WeatherService } from '../application/weather/use-cases/weather.service';
-import { GetWeatherDto } from '../application/weather/dto/get-weather.dto';
+import { WeatherService } from 'src/application/weather/use-cases/weather.service';
 import { CityValidationPipe } from 'src/common/pipes/city-validation.pipe';
+import { GetWeatherDto } from './dto/get-weather.dto';
 
 @Controller('weather')
 export class WeatherController {
